LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.meta/worker-src-none/worklet-audio.https-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'worker-src'.

Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=keep-origin&action=purge&key=1c59ba93-00bc-4ed0-9aa7-e13ac10745c1&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=no-redirect&action=purge&key=f7d16a57-b293-4e0e-87ec-d33942bf5115&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=swap-origin&action=purge&key=257cedf5-88ab-4c60-84b8-76eaa8f27452&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?action=purge&key=e3985b25-9f3f-46c9-8829-0c8751d67e7a&path=%2Fmixed-content

FAIL Content Security Policy: Expects allowed for worklet-audio to cross-https origin and keep-origin redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Cross-origin script load denied by Cross-Origin Resource Sharing policy."
PASS Content Security Policy: Expects allowed for worklet-audio to cross-https origin and keep-origin redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects allowed for worklet-audio to cross-https origin and no-redirect redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Cross-origin script load denied by Cross-Origin Resource Sharing policy."
PASS Content Security Policy: Expects allowed for worklet-audio to cross-https origin and no-redirect redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects allowed for worklet-audio to cross-https origin and swap-origin redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Cross-origin script load denied by Cross-Origin Resource Sharing policy."
PASS Content Security Policy: Expects allowed for worklet-audio to cross-https origin and swap-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and keep-origin redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and keep-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and no-redirect redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and no-redirect redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects allowed for worklet-audio to same-https origin and swap-origin redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Importing a module script is canceled."
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and swap-origin redirection from https context.: securitypolicyviolation
